+++ packages/blocks/CHANGELOG.md
@@ -1,5 +1,12 @@
 # druxt-blocks
 
+## 0.14.5
+
+### Patch Changes
+
+- Updated dependencies [60ee4e8]
+  - druxt-entity@0.23.0
+
 ## 0.14.4
 
 ### Patch Changes