+++ packages/site/CHANGELOG.md
@@ -1,5 +1,15 @@
 # druxt-site
 
+## 0.10.5
+
+### Patch Changes
+
+- Updated dependencies [60ee4e8]
+  - druxt-schema@0.10.0
+  - druxt-entity@0.23.0
+  - druxt-blocks@0.14.5
+  - druxt-views@0.17.2
+
 ## 0.10.4
 
 ### Patch Changes